p4 addition
p4 only half works. It works for the test data and returns what I believe to be a viable solution as I doubled checked it with another's code who solved the challenge. However on the data I was presented this code returns the wrong number... will work on it and maybe fix. Day 4 Part 2 is not happening.